Gui: Generalize DlgPropertyLink for objects
DlgPropertyLink is now a subclass of DlgDocumentObject for more general
use of selecting objects.
