+++ src/Gui/Dialogs/DlgPropertyLink.cpp
@@ -1,25 +1,25 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
-
-/***************************************************************************
- *   Copyright (c) 2014 Werner Mayer <wmayer[at]users.sourceforge.net>     *
- *                                                                         *
- *   This file is part of the FreeCAD CAx development system.              *
- *                                                                         *
- *   This library is free software; you can redistribute it and/or         *
- *   modify it under the terms of the GNU Library General Public           *
- *   License as published by the Free Software Foundation; either          *
- *   version 2 of the License, or (at your option) any later version.      *
- *                                                                         *
- *   This library  is distributed in the hope that it will be useful,      *
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
- *   GNU Library General Public License for more details.                  *
- *                                                                         *
- *   You should have received a copy of the GNU Library General Public     *
- *   License along with this library; see the file COPYING.LIB. If not,    *
- *   write to the Free Software Foundation, Inc., 59 Temple Place,         *
- *   Suite 330, Boston, MA  02111-1307, USA                                *
- *                                                                         *
+/****************************************************************************
+ *                                                                          *
+ *   Copyright (c) 2014 Werner Mayer <wmayer[at]users.sourceforge.net>      *
+ *   Copyright (c) 2025 Pieter Hijma <info@pieterhijma.net>                 *
+ *                                                                          *
+ *   This file is part of FreeCAD.                                          *
+ *                                                                          *
+ *   FreeCAD is free software: you can redistribute it and/or modify it     *
+ *   under the terms of the GNU Lesser General Public License as            *
+ *   published by the Free Software Foundation, either version 2.1 of the   *
+ *   License, or (at your option) any later version.                        *
+ *                                                                          *
+ *   FreeCAD is distributed in the hope that it will be useful, but         *
+ *   WITHOUT ANY WARRANTY; without even the implied warranty of             *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU       *
+ *   Lesser General Public License for more details.                        *
+ *                                                                          *
+ *   You should have received a copy of the GNU Lesser General Public       *
+ *   License along with FreeCAD. If not, see                                *
+ *   <https://www.gnu.org/licenses/>.                                       *
+ *                                                                          *
  ***************************************************************************/
 
 #include <algorithm>
@@ -36,100 +36,22 @@
 #include <Base/Tools.h>
 
 #include "Dialogs/DlgPropertyLink.h"
-#include "ui_DlgPropertyLink.h"
+#include "ui_DlgDocumentObject.h"
 #include "Application.h"
 #include "Document.h"
 #include "BitmapFactory.h"
 #include "PropertyView.h"
 #include "Selection.h"
-#include "Tree.h"
-#include "TreeParams.h"
-#include "View3DInventor.h"
 #include "ViewProviderDocumentObject.h"
 
 
 using namespace Gui::Dialog;
 
-class ItemDelegate: public QStyledItemDelegate
-{
-public:
-    explicit ItemDelegate(QObject* parent = nullptr)
-        : QStyledItemDelegate(parent)
-    {}
-
-    QWidget* createEditor(
-        QWidget* parent,
-        const QStyleOptionViewItem& option,
-        const QModelIndex& index
-    ) const override
-    {
-        if (index.column() != 1) {
-            return nullptr;
-        }
-        return QStyledItemDelegate::createEditor(parent, option, index);
-    }
-};
-
 /* TRANSLATOR Gui::Dialog::DlgPropertyLink */
 
 DlgPropertyLink::DlgPropertyLink(QWidget* parent)
-    : QDialog(parent)
-    , SelectionObserver(false, ResolveMode::NoResolve)
-    , ui(new Ui_DlgPropertyLink)
-{
-#if defined(Q_OS_MACOS)
-    // Clicking the 3D view would push this window behind the main window without this.
-    setWindowFlag(Qt::WindowStaysOnTopHint);
-#endif
-    // clang-format off
-    ui->setupUi(this);
-    connect(ui->checkObjectType, &QCheckBox::toggled,
-            this, &DlgPropertyLink::onObjectTypeToggled);
-    connect(ui->typeTree, &QTreeWidget::itemSelectionChanged,
-            this, &DlgPropertyLink::onTypeTreeItemSelectionChanged);
-    connect(ui->searchBox, &ExpressionLineEdit::textChanged,
-            this, &DlgPropertyLink::onSearchBoxTextChanged);
-
-    ui->typeTree->hide();
-    ui->searchBox->installEventFilter(this);
-    ui->searchBox->setNoProperty(true);
-    ui->searchBox->setExactMatch(Gui::ExpressionParameter::instance()->isExactMatch());
-
-    timer = new QTimer(this);
-    timer->setSingleShot(true);
-
-    connect(timer, &QTimer::timeout, this, &DlgPropertyLink::onTimer);
-
-    ui->treeWidget->setEditTriggers(QAbstractItemView::DoubleClicked);
-    ui->treeWidget->setItemDelegate(new ItemDelegate(this));
-    ui->treeWidget->setMouseTracking(true);
-    connect(ui->treeWidget, &QTreeWidget::itemEntered,
-            this, &DlgPropertyLink::onItemEntered);
-
-    connect(ui->treeWidget, &QTreeWidget::itemExpanded,
-            this, &DlgPropertyLink::onItemExpanded);
-
-    connect(ui->treeWidget, &QTreeWidget::itemSelectionChanged, this, &DlgPropertyLink::onItemSelectionChanged);
-
-    connect(ui->searchBox, &QLineEdit::returnPressed, this, &DlgPropertyLink::onItemSearch);
-
-    connect(ui->buttonBox, &QDialogButtonBox::clicked, this, &DlgPropertyLink::onClicked);
-
-    refreshButton = ui->buttonBox->addButton(tr("Reset"), QDialogButtonBox::ActionRole);
-    resetButton = ui->buttonBox->addButton(tr("Clear"), QDialogButtonBox::ResetRole);
-    // clang-format on
-}
-
-/**
- *  Destroys the object and frees any allocated resources
- */
-DlgPropertyLink::~DlgPropertyLink()
-{
-    detachObserver();
-
-    // no need to delete child widgets, Qt does it all for us
-    delete ui;
-}
+    : DlgDocumentObject(parent)
+{}
 
 QList<App::SubObjectT> DlgPropertyLink::getLinksFromProperty(const App::PropertyLinkBase* prop)
 {
@@ -257,6 +179,7 @@ QString DlgPropertyLink::formatLinks(App::Document* ownerDoc, QList<App::SubObje
 
 void DlgPropertyLink::init(const App::DocumentObjectT& prop, bool tryFilter)
 {
+    setWindowTitle(tr("Link"));
     ui->treeWidget->blockSignals(true);
     ui->treeWidget->clear();
     ui->treeWidget->blockSignals(false);
@@ -277,7 +200,7 @@ void DlgPropertyLink::init(const App::DocumentObjectT& prop, bool tryFilter)
     selections.clear();
 
     objProp = prop;
-    auto owner = objProp.getObject();
+    owner = objProp.getObject();
     if (!owner || !owner->isAttachedToDocument()) {
         return;
     }
@@ -367,8 +290,8 @@ void DlgPropertyLink::init(const App::DocumentObjectT& prop, bool tryFilter)
         auto item = new QTreeWidgetItem(ui->treeWidget);
         item->setIcon(0, docIcon);
         item->setText(0, QString::fromUtf8(d->Label.getValue()));
-        item->setData(0, Qt::UserRole, QByteArray(""));
-        item->setData(0, Qt::UserRole + 1, QByteArray(d->getName()));
+        item->setData(0, ObjectNameRole, QByteArray(""));
+        item->setData(0, DocNameRole, QByteArray(d->getName()));
         item->setFlags(Qt::ItemIsEnabled);
         item->setChildIndicatorPolicy(QTreeWidgetItem::ShowIndicator);
         if (expandDocs.count(d)) {
@@ -479,391 +402,9 @@ void DlgPropertyLink::onClicked(QAbstractButton* button)
     }
 }
 
-void DlgPropertyLink::hideEvent(QHideEvent* ev)
-{
-    detachObserver();
-    QDialog::hideEvent(ev);
-}
-
-void DlgPropertyLink::closeEvent(QCloseEvent* ev)
-{
-    detachObserver();
-    QDialog::closeEvent(ev);
-}
-
-void DlgPropertyLink::attachObserver()
-{
-    if (isSelectionAttached()) {
-        return;
-    }
-
-    Gui::Selection().selStackPush();
-    attachSelection();
-
-    if (!parentView) {
-        for (auto p = parent(); p; p = p->parent()) {
-            auto view = qobject_cast<Gui::PropertyView*>(p);
-            if (view) {
-                parentView = view;
-                for (auto& sel : Gui::Selection().getCompleteSelection(ResolveMode::NoResolve)) {
-                    savedSelections.emplace_back(sel.DocName, sel.FeatName, sel.SubName);
-                }
-                break;
-            }
-        }
-    }
-    auto view = qobject_cast<Gui::PropertyView*>(parentView.data());
-    if (view) {
-        view->blockSelection(true);
-    }
-}
-
-void DlgPropertyLink::showEvent(QShowEvent* ev)
-{
-    attachObserver();
-    QDialog::showEvent(ev);
-}
-
-void DlgPropertyLink::onItemEntered(QTreeWidgetItem*)
-{
-    int timeout = Gui::TreeParams::getPreSelectionDelay() / 2;
-    if (timeout < 0) {
-        timeout = 1;
-    }
-    timer->start(timeout);
-    Gui::Selection().rmvPreselect();
-}
-
-void DlgPropertyLink::leaveEvent(QEvent* ev)
-{
-    Gui::Selection().rmvPreselect();
-    QDialog::leaveEvent(ev);
-}
-
-void DlgPropertyLink::detachObserver()
-{
-    if (isSelectionAttached()) {
-        detachSelection();
-    }
-
-    auto view = qobject_cast<Gui::PropertyView*>(parentView.data());
-    if (view && !savedSelections.empty()) {
-        try {
-            Gui::Selection().clearSelection();
-        }
-        catch (Py::Exception& e) {
-            e.clear();
-        }
-        for (auto& sel : savedSelections) {
-            if (sel.getSubObject()) {
-                Gui::Selection().addSelection(
-                    sel.getDocumentName().c_str(),
-                    sel.getObjectName().c_str(),
-                    sel.getSubName().c_str()
-                );
-            }
-        }
-        savedSelections.clear();
-    }
-    if (view) {
-        view->blockSelection(false);
-    }
-
-    parentView = nullptr;
-}
-
-void DlgPropertyLink::onItemSelectionChanged()
-{
-    auto newSelections = ui->treeWidget->selectedItems();
-
-    if (newSelections.isEmpty() || selections.contains(newSelections.back())) {
-        selections = newSelections;
-        if (newSelections.isEmpty()) {
-            currentObj = nullptr;
-        }
-        return;
-    }
-
-    selections = newSelections;
-
-    auto sobjs = getLinkFromItem(newSelections.back());
-    App::DocumentObject* obj = !sobjs.empty() ? sobjs.front().getObject() : nullptr;
-    if (!obj) {
-        Gui::Selection().clearSelection();
-        return;
-    }
-
-    bool focus = false;
-    // Do auto view switch if tree view does not do it
-    if (!TreeParams::getSyncView()) {
-        focus = ui->treeWidget->hasFocus();
-        auto doc = Gui::Application::Instance->getDocument(sobjs.front().getDocumentName().c_str());
-        if (doc) {
-            auto vp = freecad_cast<Gui::ViewProviderDocumentObject*>(doc->getViewProvider(obj));
-            if (vp) {
-                // If the view provider uses a special window for rendering, switch to it
-                MDIView* view = vp->getMDIView();
-                if (view) {
-                    doc->setActiveWindow(view);
-                }
-                else {
-                    doc->setActiveView(vp, Gui::View3DInventor::getClassTypeId());
-                }
-            }
-        }
-    }
-
-    // Sync 3d view selection. To give a better visual feedback, we
-    // only keep the latest selection.
-    bool blocked = blockSelection(true);
-    Gui::Selection().clearSelection();
-    for (auto& sobj : sobjs) {
-        Gui::Selection().addSelection(
-            sobj.getDocumentName().c_str(),
-            sobj.getObjectName().c_str(),
-            sobj.getSubName().c_str()
-        );
-    }
-    blockSelection(blocked);
-
-    // Enforce single parent
-    if (singleParent && currentObj && currentObj != obj) {
-        ui->treeWidget->blockSignals(true);
-        const auto items = ui->treeWidget->selectedItems();
-        for (auto item : items) {
-            if (item != selections.back()) {
-                item->setSelected(false);
-            }
-        }
-        auto last = selections.back();
-        selections.clear();
-        selections.append(last);
-        ui->treeWidget->blockSignals(false);
-    }
-    currentObj = obj;
-
-    if (focus) {
-        // FIXME: does not work, why?
-        ui->treeWidget->setFocus();
-    }
-}
-
-QTreeWidgetItem* DlgPropertyLink::findItem(App::DocumentObject* obj, const char* subname, bool* pfound)
-{
-    if (pfound) {
-        *pfound = false;
-    }
-
-    if (!obj || !obj->isAttachedToDocument()) {
-        return nullptr;
-    }
-
-    std::vector<App::DocumentObject*> sobjs;
-    if (!Base::Tools::isNullOrEmpty(subname)) {
-        if (!allowSubObject) {
-            obj = obj->getSubObject(subname);
-            if (!obj) {
-                return nullptr;
-            }
-        }
-        else {
-            sobjs = obj->getSubObjectList(subname);
-        }
-    }
-
-    auto itDoc = docItems.find(obj->getDocument());
-    if (itDoc == docItems.end()) {
-        return nullptr;
-    }
-    onItemExpanded(itDoc->second);
-
-    auto it = itemMap.find(obj);
-    if (it == itemMap.end() || it->second->isHidden()) {
-        return nullptr;
-    }
-
-    if (!allowSubObject) {
-        if (pfound) {
-            *pfound = true;
-        }
-        return it->second;
-    }
-
-    QTreeWidgetItem* item = it->second;
-
-    bool first = true;
-    for (auto o : sobjs) {
-        if (first) {
-            first = false;
-            continue;
-        }
-        onItemExpanded(item);
-        bool found = false;
-        for (int i = 0, count = item->childCount(); i < count; ++i) {
-            auto child = item->child(i);
-            if (strcmp(o->getNameInDocument(), child->data(0, Qt::UserRole).toByteArray().constData())
-                == 0) {
-                item = child;
-                found = true;
-                break;
-            }
-        }
-        if (!found) {
-            return item;
-        }
-    }
-    if (pfound) {
-        *pfound = true;
-    }
-    return item;
-}
-
-void DlgPropertyLink::onSelectionChanged(const Gui::SelectionChanges& msg)
-{
-    if (msg.Type != SelectionChanges::AddSelection) {
-        return;
-    }
-
-    bool found = false;
-    auto selObj = msg.Object.getObject();
-
-    App::ElementNamePair elementName;
-    const char* subname = msg.pSubName;
-    if (!ui->checkSubObject->isChecked()) {
-        selObj = App::GeoFeature::resolveElement(selObj, subname, elementName);
-        if (!selObj) {
-            return;
-        }
-        subname = elementName.oldName.c_str();
-    }
-
-    auto item = findItem(selObj, msg.pSubName, &found);
-    if (!item) {
-        return;
-    }
-
-    if (!item->isSelected()) {
-        ui->treeWidget->blockSignals(true);
-        if (singleSelect || (singleParent && currentObj && currentObj != selObj)) {
-            ui->treeWidget->selectionModel()->clearSelection();
-        }
-        currentObj = selObj;
-        item->setSelected(true);
-        selections.append(item);
-        ui->treeWidget->blockSignals(false);
-    }
-
-    ui->treeWidget->scrollToItem(item);
-    if (allowSubObject) {
-        QString element = QString::fromUtf8(msg.Object.getOldElementName().c_str());
-        if (element.size()) {
-            QStringList list;
-            QString text = item->text(1);
-            if (text.size()) {
-                list = text.split(QLatin1Char(','));
-            }
-            if (list.indexOf(element) < 0) {
-                list << element;
-                item->setText(1, list.join(QLatin1String(",")));
-                subSelections.insert(item);
-            }
-        }
-        else if (subSelections.erase(item)) {
-            item->setText(1, QString());
-        }
-    }
-}
-
-void DlgPropertyLink::accept()
-{
-    QDialog::accept();
-}
-
-static QTreeWidgetItem* _getLinkFromItem(std::ostringstream& ss, QTreeWidgetItem* item, const char* objName)
-{
-    auto parent = item->parent();
-    assert(parent);
-    QByteArray nextName = parent->data(0, Qt::UserRole).toByteArray();
-    if (nextName.isEmpty()) {
-        return item;
-    }
-
-    item = _getLinkFromItem(ss, parent, nextName);
-    ss << objName << '.';
-    return item;
-}
-
-QList<App::SubObjectT> DlgPropertyLink::getLinkFromItem(QTreeWidgetItem* item, bool needSubName) const
-{
-    QList<App::SubObjectT> res;
-
-    auto parent = item->parent();
-    if (!parent) {
-        return res;
-    }
-
-    std::ostringstream ss;
-    auto parentItem = _getLinkFromItem(ss, item, item->data(0, Qt::UserRole).toByteArray().constData());
-
-    App::SubObjectT sobj(
-        parentItem->data(0, Qt::UserRole + 1).toByteArray().constData(),
-        parentItem->data(0, Qt::UserRole).toByteArray().constData(),
-        ss.str().c_str()
-    );
-
-    QString elements;
-    if (needSubName && allowSubObject) {
-        elements = item->text(1);
-    }
-
-    if (elements.isEmpty()) {
-        res.append(App::SubObjectT());
-        res.last() = std::move(sobj);
-        return res;
-    }
-
-    const auto split = elements.split(QLatin1Char(','));
-    for (const QString& element : split) {
-        res.append(App::SubObjectT());
-        res.last() = App::SubObjectT(
-            sobj.getDocumentName().c_str(),
-            sobj.getObjectName().c_str(),
-            (sobj.getSubName() + element.toLatin1().constData()).c_str()
-        );
-    }
-    return res;
-}
-
-void DlgPropertyLink::onTimer()
-{
-    auto item = ui->treeWidget->itemAt(ui->treeWidget->viewport()->mapFromGlobal(QCursor::pos()));
-    if (!item) {
-        return;
-    }
-    auto sobjs = getLinkFromItem(item);
-    if (sobjs.isEmpty()) {
-        return;
-    }
-    const auto& sobj = sobjs.front();
-    Gui::Selection().setPreselect(
-        sobj.getDocumentName().c_str(),
-        sobj.getObjectName().c_str(),
-        sobj.getSubName().c_str(),
-        0,
-        0,
-        0,
-        Gui::SelectionChanges::MsgSource::TreeView
-    );
-}
-
 QList<App::SubObjectT> DlgPropertyLink::currentLinks() const
 {
-    auto items = ui->treeWidget->selectedItems();
-    QList<App::SubObjectT> res;
-    for (auto item : items) {
-        res.append(getLinkFromItem(item));
-    }
-    return res;
+    return currentSubObjects();
 }
 
 QList<App::SubObjectT> DlgPropertyLink::originalLinks() const
@@ -902,351 +443,6 @@ QString DlgPropertyLink::linksToPython(const QList<App::SubObjectT>& links)
     }
 
     return QString::fromUtf8(ss.str().c_str());
-}
-
-void DlgPropertyLink::filterObjects()
-{
-    for (int i = 0, count = ui->treeWidget->topLevelItemCount(); i < count; ++i) {
-        auto item = ui->treeWidget->topLevelItem(i);
-        for (int j = 0, c = item->childCount(); j < c; ++j) {
-            filterItem(item->child(j));
-        }
-    }
-}
-
-void DlgPropertyLink::filterItem(QTreeWidgetItem* item)
-{
-    if (filterType(item)) {
-        item->setHidden(true);
-        return;
-    }
-    item->setHidden(false);
-    for (int i = 0, count = item->childCount(); i < count; ++i) {
-        filterItem(item->child(i));
-    }
-}
-
-bool DlgPropertyLink::eventFilter(QObject* obj, QEvent* e)
-{
-    if (obj == ui->searchBox && e->type() == QEvent::KeyPress
-        && static_cast<QKeyEvent*>(e)->key() == Qt::Key_Escape) {
-        ui->searchBox->setText(QString());
-        return true;
-    }
-    return QDialog::eventFilter(obj, e);
-}
-
-void DlgPropertyLink::onItemSearch()
-{
-    itemSearch(ui->searchBox->text(), true);
-}
-
-void DlgPropertyLink::keyPressEvent(QKeyEvent* ev)
-{
-    if (ev->key() == Qt::Key_Enter || ev->key() == Qt::Key_Return) {
-        if (ui->searchBox->hasFocus()) {
-            return;
-        }
-    }
-    QDialog::keyPressEvent(ev);
-}
-
-void DlgPropertyLink::itemSearch(const QString& text, bool select)
-{
-    if (searchItem) {
-        searchItem->setBackground(0, bgBrush);
-    }
-
-    auto owner = objProp.getObject();
-    if (!owner) {
-        return;
-    }
-
-    std::string txt(text.toUtf8().constData());
-    try {
-        if (txt.empty()) {
-            return;
-        }
-        if (txt.find("<<") == std::string::npos) {
-            auto pos = txt.find('.');
-            if (pos == std::string::npos) {
-                txt += '.';
-            }
-            else if (pos != txt.size() - 1) {
-                txt.insert(pos + 1, "<<");
-                if (txt.back() != '.') {
-                    txt += '.';
-                }
-                txt += ">>.";
-            }
-        }
-        else if (txt.back() != '.') {
-            txt += '.';
-        }
-        txt += "_self";
-        auto path = App::ObjectIdentifier::parse(owner, txt);
-        if (path.getPropertyName() != "_self") {
-            return;
-        }
-
-        App::DocumentObject* obj = path.getDocumentObject();
-        if (!obj) {
-            return;
-        }
-
-        bool found;
-        const char* subname = path.getSubObjectName().c_str();
-        QTreeWidgetItem* item = findItem(obj, subname, &found);
-        if (!item) {
-            return;
-        }
-
-        if (select) {
-            if (!found) {
-                return;
-            }
-            Gui::Selection()
-                .addSelection(obj->getDocument()->getName(), obj->getNameInDocument(), subname);
-        }
-        else {
-            Selection().setPreselect(
-                obj->getDocument()->getName(),
-                obj->getNameInDocument(),
-                subname,
-                0,
-                0,
-                0,
-                Gui::SelectionChanges::MsgSource::TreeView
-            );
-            searchItem = item;
-            ui->treeWidget->scrollToItem(searchItem);
-            bgBrush = searchItem->background(0);
-            searchItem->setBackground(0, QColor(255, 255, 0, 100));
-        }
-    }
-    catch (...) {
-    }
-}
-
-QTreeWidgetItem* DlgPropertyLink::createItem(App::DocumentObject* obj, QTreeWidgetItem* parent)
-{
-    if (!obj || !obj->isAttachedToDocument()) {
-        return nullptr;
-    }
-
-    if (inList.find(obj) != inList.end()) {
-        return nullptr;
-    }
-
-    auto vp = freecad_cast<ViewProviderDocumentObject*>(Application::Instance->getViewProvider(obj));
-    if (!vp) {
-        return nullptr;
-    }
-
-    QTreeWidgetItem* item;
-    if (parent) {
-        item = new QTreeWidgetItem(parent);
-    }
-    else {
-        item = new QTreeWidgetItem(ui->treeWidget);
-    }
-    item->setIcon(0, vp->getIcon());
-    item->setText(0, QString::fromUtf8((obj)->Label.getValue()));
-    item->setData(0, Qt::UserRole, QByteArray(obj->getNameInDocument()));
-    item->setData(0, Qt::UserRole + 1, QByteArray(obj->getDocument()->getName()));
-
-    if (allowSubObject) {
-        item->setChildIndicatorPolicy(
-            !obj->getLinkedObject(true)->getOutList().empty() ? QTreeWidgetItem::ShowIndicator
-                                                              : QTreeWidgetItem::DontShowIndicator
-        );
-        item->setFlags(item->flags() | Qt::ItemIsEditable | Qt::ItemIsUserCheckable);
-    }
-
-    const auto typeName = obj->getTypeId().getName();
-    auto typeData = QByteArray::fromRawData(typeName.data(), typeName.size());
-    item->setData(0, Qt::UserRole + 2, typeData);
-
-    QByteArray proxyType;
-    auto prop = freecad_cast<App::PropertyPythonObject*>(obj->getPropertyByName("Proxy"));
-    if (prop) {
-        Base::PyGILStateLocker lock;
-        Py::Object proxy = prop->getValue();
-        if (!proxy.isNone() && !proxy.isString()) {
-            const char* name = nullptr;
-            if (proxy.hasAttr("__class__")) {
-                proxyType = QByteArray::fromStdString(proxy.getAttr("__class__").as_string());
-            }
-            else {
-                name = proxy.ptr()->ob_type->tp_name;
-                proxyType = QByteArray::fromRawData(name, strlen(name));
-            }
-            auto it = typeItems.find(proxyType);
-            if (it != typeItems.end()) {
-                proxyType = it->first;
-            }
-            else if (name) {
-                proxyType = QByteArray(name, proxyType.size());
-            }
-        }
-    }
-    item->setData(0, Qt::UserRole + 3, proxyType);
-
-    filterItem(item);
-    return item;
-}
-
-QTreeWidgetItem* DlgPropertyLink::createTypeItem(Base::Type type)
-{
-    if (type.isBad()) {
-        return nullptr;
-    }
-
-    QTreeWidgetItem* item = nullptr;
-    if (!type.isBad() && type != App::DocumentObject::getClassTypeId()) {
-        Base::Type parentType = type.getParent();
-        if (!parentType.isBad()) {
-            const auto name = parentType.getName();
-            QByteArray typeData(name.data(), name.size());
-            auto& typeItem = typeItems[typeData];
-            if (!typeItem) {
-                typeItem = createTypeItem(parentType);
-                typeItem->setData(0, Qt::UserRole, typeData);
-            }
-            item = typeItem;
-        }
-    }
-
-    if (!item) {
-        item = new QTreeWidgetItem(ui->typeTree);
-    }
-    else {
-        item = new QTreeWidgetItem(item);
-    }
-    item->setExpanded(true);
-    item->setText(0, QString::fromUtf8(type.getName()));
-    if (type == App::DocumentObject::getClassTypeId()) {
-        item->setFlags(Qt::ItemIsEnabled);
-    }
-    return item;
-}
-
-bool DlgPropertyLink::filterType(QTreeWidgetItem* item)
-{
-    auto proxyType = item->data(0, Qt::UserRole + 3).toByteArray();
-    QTreeWidgetItem* proxyItem = nullptr;
-    if (proxyType.size()) {
-        auto& pitem = typeItems[proxyType];
-        if (!pitem) {
-            pitem = new QTreeWidgetItem(ui->typeTree);
-            pitem->setText(0, QString::fromUtf8(proxyType));
-            pitem->setIcon(0, item->icon(0));
-            pitem->setData(0, Qt::UserRole, proxyType);
-        }
-        proxyItem = pitem;
-    }
-
-    auto typeData = item->data(0, Qt::UserRole + 2).toByteArray();
-    Base::Type type = Base::Type::fromName(std::string_view(typeData.data(), typeData.size()));
-    if (type.isBad()) {
-        return false;
-    }
-
-    QTreeWidgetItem*& typeItem = typeItems[typeData];
-    if (!typeItem) {
-        typeItem = createTypeItem(type);
-        typeItem->setData(0, Qt::UserRole, typeData);
-    }
-
-    if (!proxyType.size()) {
-        QIcon icon = typeItem->icon(0);
-        if (icon.isNull()) {
-            typeItem->setIcon(0, item->icon(0));
-        }
-    }
-
-    if (!ui->checkObjectType->isChecked() || selectedTypes.empty()) {
-        return false;
-    }
-
-    if (proxyItem && selectedTypes.count(proxyType)) {
-        return false;
-    }
-
-    for (auto t = type; !t.isBad() && t != App::DocumentObject::getClassTypeId(); t = t.getParent()) {
-        const auto name = t.getName();
-        if (selectedTypes.count(QByteArray(name.data(), name.size()))) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-void DlgPropertyLink::onItemExpanded(QTreeWidgetItem* item)
-{
-    if (item->childCount()) {
-        return;
-    }
-
-    QByteArray docName = item->data(0, Qt::UserRole + 1).toByteArray();
-    auto doc = App::GetApplication().getDocument(docName);
-    if (!doc) {
-        return;
-    }
-
-    QByteArray objName = item->data(0, Qt::UserRole).toByteArray();
-    if (objName.isEmpty()) {
-        for (auto obj : doc->getObjects()) {
-            auto newItem = createItem(obj, item);
-            if (newItem) {
-                itemMap[obj] = newItem;
-            }
-        }
-    }
-    else if (allowSubObject) {
-        auto obj = doc->getObject(objName);
-        if (!obj) {
-            return;
-        }
-        std::set<App::DocumentObject*> childSet;
-        std::string sub;
-        for (auto child : obj->getLinkedObject(true)->getOutList()) {
-            if (!childSet.insert(child).second) {
-                continue;
-            }
-            sub = child->getNameInDocument();
-            sub += ".";
-            if (obj->getSubObject(sub.c_str())) {
-                createItem(child, item);
-            }
-        }
-    }
-}
-
-void DlgPropertyLink::onObjectTypeToggled(bool on)
-{
-    ui->typeTree->setVisible(on);
-    filterObjects();
-}
-
-void DlgPropertyLink::onTypeTreeItemSelectionChanged()
-{
-
-    selectedTypes.clear();
-    const auto items = ui->typeTree->selectedItems();
-    for (auto item : items) {
-        selectedTypes.insert(item->data(0, Qt::UserRole).toByteArray());
-    }
-
-    if (ui->checkObjectType->isChecked()) {
-        filterObjects();
-    }
-}
-
-void DlgPropertyLink::onSearchBoxTextChanged(const QString& text)
-{
-    itemSearch(text, false);
 }
 
 #include "moc_DlgPropertyLink.cpp"
